#84 招待ボタンをコンポーネント化
