+++ src/pages/Lobby.tsx
@@ -6,34 +6,13 @@ import Grid from "@mui/material/Grid";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import SendIcon from "@mui/icons-material/Send";
-import PersonAddAlt1Icon from "@mui/icons-material/PersonAddAlt1";
 import Typography from "@mui/material/Typography";
-import Dialog from "@mui/material/Dialog";
-import DialogActions from "@mui/material/DialogActions";
-import DialogContent from "@mui/material/DialogContent";
-import DialogContentText from "@mui/material/DialogContentText";
-import DialogTitle from "@mui/material/DialogTitle";
-import IconButton from "@mui/material/IconButton";
-import CloseIcon from "@mui/icons-material/Close";
-import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
-import { CopyToClipboard } from "react-copy-to-clipboard";
-import { TwitterShareButton, TwitterIcon, LineShareButton, LineIcon } from "react-share";
-import ContentCopyIcon from "@mui/icons-material/ContentCopy";
 import theme from "../styles";
 import UserList from "../components/UserList";
+import InviteButton from "../components/InviteButton";
 
 const Lobby = () => {
-  const [open, setOpen] = useState(false);
-  const url = "http://localhost:3000/lobby";
-
-  const handleClickOpen = () => {
-    setOpen(true);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
+  const url = "oekaki-app/lobby";
   return (
     <ThemeProvider theme={theme}>
       <CssBaseline />
@@ -97,51 +76,7 @@ const Lobby = () => {
           </Box>
           <Box display="flex">
             <Box sx={{ pr: 1 }}>
-              <Button
-                variant="contained"
-                onClick={handleClickOpen}
-                size="large"
-                endIcon={<PersonAddAlt1Icon />}
-                sx={{
-                  background: theme.palette.secondary.main,
-                  "&:hover": {
-                    background: theme.palette.secondary.dark,
-                  },
-                }}
-              >
-                招待
-              </Button>
-              <Dialog open={open} onClose={handleClose}>
-                <Box display="flex" justifyContent="space-between">
-                  <DialogTitle>招待</DialogTitle>
-                  <DialogActions>
-                    <IconButton aria-label="close" onClick={handleClose}>
-                      <CloseIcon />
-                    </IconButton>
-                  </DialogActions>
-                </Box>
-                <DialogContent>
-                  <DialogContentText>下のリンクをコピーして共有しよう！</DialogContentText>
-                  <Box sx={{ pt: 4 }}>
-                    <CopyToClipboard text={url}>
-                      <IconButton size="small">
-                        <ContentCopyIcon />
-                      </IconButton>
-                    </CopyToClipboard>
-                    <SyntaxHighlighter>{url}</SyntaxHighlighter>
-                  </Box>
-                  <Box display="flex" justifyContent="flex-end">
-                    <Box sx={{ pr: 1 }}>
-                      <TwitterShareButton url={url} title="みんなで遊ぼう、お絵かきアプリ！" hashtags={["oekakiapp"]}>
-                        <TwitterIcon size={30} round />
-                      </TwitterShareButton>
-                    </Box>
-                    <LineShareButton url={url} title="みんなで遊ぼう、お絵かきアプリ！">
-                      <LineIcon size={30} round />
-                    </LineShareButton>
-                  </Box>
-                </DialogContent>
-              </Dialog>
+              <InviteButton url={url} />
             </Box>
             <Link to="/main" style={{ textDecoration: "none" }}>
               <Box sx={{ pl: 1 }}>
